% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresDefaultLoader.R
\name{features.defaultLoader}
\alias{features.defaultLoader}
\title{The Default Loader for Features}
\usage{
features.defaultLoader(featureFolder, components)
}
\arguments{
\item{featureFolder}{the folder with the features}

\item{components}{the components}
}
\value{
a named list, where the names are the feature names and the values
  the corresponding feature values
}
\description{
The default loader for the model features is the
  \code{\link{features.txtLoader}}
}
