% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresTxtLoader.R
\name{features.txtLoader}
\alias{features.txtLoader}
\title{The Text-based Loader for Features}
\usage{
features.txtLoader(featureFolder, components, values.eval = TRUE)
}
\arguments{
\item{featureFolder}{the folder with the features}

\item{components}{the components}

\item{values.eval}{the evaluated values}
}
\value{
a named list, where the names are the feature names and the values
  the corresponding feature values
}
\description{
The text-based loader for the model features searches the
  \code{featureFolder} for all text files whose names match to the supplied
  path \code{components}. Each such file is loaded a Java-style properties
  file with \code{name=value} style and \code{#}-based comments. The
  resulting list contains all the corresponding names and values. Values can
  be parsed if \code{value.eval} is set to \code{TRUE}.
}
