% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\docType{class}
\name{Models-class}
\alias{Models-class}
\alias{Models}
\title{A List of Models (\code{\link{RegressionResult}}s) for a Specific Setup
  Identified by a Name and a List of Features}
\description{
Each instance of this  class represents list of regression
  models for a specific setup identified by a list of features and a name.
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of this setup}

\item{\code{features}}{a list of name-value pairs denoting the features of the setup}

\item{\code{models}}{the list of instances of \code{\link{RegressionResult}} for
this setup}
}}

