% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLearn.R
\name{Models.batchLearn}
\alias{Models.batchLearn}
\title{Batch-Learn Regression Models for Process Mining}
\usage{
Models.batchLearn(learners = .default, cores = detectCores(), ...)
}
\arguments{
\item{learners}{the model learners to be applied}

\item{cores}{the number of cores to use (\code{>1L} leads to parallel
execution)}

\item{...}{Arguments passed on to \code{regressoR::regressoR.batchLearn}
\describe{
  \item{source}{the source directory, which is recursively searched for files
with data to be modeled}
  \item{destination}{the destination folder, will be created if not existing}
  \item{loader}{a loader function which accepts a vector of paths and is
supposed to return an \code{n*2} matrix where the first column contains the
\code{x} values and the second column the \code{y} values to model}
  \item{selector}{a regular expression against which file names are matched.
Only matching files are considered.}
  \item{check.directory}{a function receiving a \code{root} folder and the
\code{path} under \code{root} and decides whether this directory path
should be recursively investigated (be default always \code{TRUE})}
  \item{learn.single}{should every single file matching to the \code{selector}
be modeled separately ?}
  \item{learn.all}{should all the files in one directory combined and modeled
at once?}
  \item{representations}{a function which can transform a \code{x}/\code{y}
dataset into a set of transformed data instances}
  \item{metricGenerator}{the learning quality metric generator}
  \item{suffix.single}{the suffix to append to the files containing the single
models}
  \item{suffix.all}{the suffix to be applied to the files containing the models
of all files in a folder}
  \item{q}{the modelling quality parameter}
  \item{includeMetric}{should the metric used for learning be stored in the
files}
  \item{logging}{should progress information be printed: either \code{TRUE} for
printing to the console via \code{\link{print}}, \code{FALSE} for no
logging, or a path to a file receiving logging information}
  \item{returnResults}{should we return the computed results or not?}
  \item{skipExisting}{should already existing models (resulting from a
previous, incomplete execution) simply be skipped or overwritting}
}}
}
\description{
This is a wrapper around
  \code{\link[regressoR]{regressoR.batchLearn}} which should provide a
  shortcut to the recommended modeling techniques.
}
