% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLoad.R
\name{Models.batchLoad}
\alias{Models.batchLoad}
\title{Load the Models Generated by \code{\link{regressoR.batchLearn}} as \code{\link[dataManageR]{dataset}}s}
\usage{
Models.batchLoad(path = getwd(), selector = path.extensionRegExp(extensions
  = "model", before.extension = "_single"), featuresFolder = file.path(path,
  "../features"), featuresLoader = datasets.feature.load.default,
  nameProcessor = identity, check.directory = NULL, cores = 1L,
  logging = (cores <= 1L))
}
\arguments{
\item{path}{the source folder}

\item{selector}{the selector for the files}

\item{featuresFolder}{the folder where to look for feature information,
\code{NULL} if no feature information is needed}

\item{featuresLoader}{a function accepting two parameters,
\code{featureFolder} and \code{components} and returning a named list of
features, or \code{NULL} if no features are needed. see
\code{\link{features.defaultLoader}} for documentation}

\item{nameProcessor}{a function which receives a relativized path to the
folder with the current model(s) and returns a name for the models}

\item{check.directory}{a function which can choose if a directory should be
followed or not}

\item{cores}{the number of cores to be used for loading}

\item{logging}{should progress information be printed: either \code{TRUE} for
printing to the console via \code{\link{print}}, \code{FALSE} for no
logging, or a path to a file receiving logging information}
}
\value{
a list of \code{\link[dataManageR]{dataset}} instances whose \code{data} elements correspond to models
}
\description{
Load all the setup results built by
  \code{\link{regressoR.batchLearn}} from a folder structure.
}
