% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureNames.R
\name{Models.featureNames}
\alias{Models.featureNames}
\title{Get the Names of All Features of All Models}
\usage{
Models.featureNames(models)
}
\arguments{
\item{models}{the list of models from where we want to extract the feature
names}
}
\value{
the feature names (alphabetically sorted)
}
\description{
Get the names of all features of all models.
}
