% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureValues.R
\name{Models.featureValues}
\alias{Models.featureValues}
\title{Get a Vector with the Values of one Feature for all Models}
\usage{
Models.featureValues(models, featureName)
}
\arguments{
\item{models}{the list of models from where we want to extract the feature
names}

\item{featureName}{the name of the feature}
}
\value{
the vector with the values of this feature
}
\description{
For each model, get the value of the specified feature. If the
  feature is undefined, its value is \code{NA}.
}
