% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinNames.R
\name{Models.joinNames}
\alias{Models.joinNames}
\title{Join a Set of Names by Keeping only the Identical Parts}
\usage{
Models.joinNames(names, features = NULL)
}
\arguments{
\item{names}{the vector of names}

\item{features}{the vector of feature values}
}
\value{
a single string of the common name components only
}
\description{
This function receives a vector of names and tries to keep only
  their common components. Name components are separated by slashes ("/").
  If no names can be found, use the features instead.
}
