% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNamesUnique.R
\name{Models.make.names.unique}
\alias{Models.make.names.unique}
\title{Receive a List of \code{\link{Models}} Instances and Make their Names
  Unique}
\usage{
Models.make.names.unique(models)
}
\arguments{
\item{models}{the list of \code{\link{Models}} instances}
}
\value{
a list of \code{\link{Models}} instances where no name appears twice
}
\description{
Receive a list of instances of class \code{\link{Models}} and
  make sure that all of their names are unique.
}
