% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeModelsBySelection.R
\name{Models.merge.by.selection}
\alias{Models.merge.by.selection}
\title{Create a Combination or Selection of Models}
\usage{
Models.merge.by.selection(models, selection = seq_along(models),
  namer = Models.joinNames)
}
\arguments{
\item{models}{the list of instances of \code{\link{Models}}}

\item{selection}{a list of vectors of model indexes. Models in the same index are joined}

\item{namer}{a function used to join names of model sets that now have
identical features, see \code{\link{Models.joinNames}} for
documentation}
}
\value{
a list of  instances of \code{\link{Models}} where
  all names have been transformed according to \code{namer} and the features
  have been selected according to selection
}
\description{
Merge a set of models according to a given selection.
}
