% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{Models.names}
\alias{Models.names}
\title{Get the Names of All Models in a List}
\usage{
Models.names(models)
}
\arguments{
\item{models}{the list of models}
}
\value{
the vector of names of the models
}
\description{
This function extracts all the names of the models in a list of
  models and returns them in the same order as the models appear.
}
