% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{Models.new}
\alias{Models.new}
\title{Create an Instance of \code{\link{Models}}}
\usage{
Models.new(name, features, models)
}
\arguments{
\item{name}{the name of this setup}

\item{features}{a list of name-value pairs denoting the features of the setup}

\item{models}{the list of instances of \code{\link{RegressionResult}} for
this setup}
}
\description{
Instantiate \code{\link{Models}} using this method.
}
