% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLoad.R
\name{processMineR.batchLoad}
\alias{processMineR.batchLoad}
\title{Load the \code{\link{Models}} Generated by
  \code{\link{regressoR.batchLearn}}}
\usage{
processMineR.batchLoad(source = getwd(),
  selector = path.extensionRegExp(extensions = "model", before.extension =
  "_single"), featuresFolder = file.path(source, "../features"),
  featuresLoader = features.defaultLoader, nameProcessor = identity,
  check.directory = NULL, cores = 1L, logging = (cores <= 1L))
}
\arguments{
\item{source}{the source folder}

\item{selector}{the selector for the files}

\item{featuresFolder}{the folder where to look for feature information,
\code{NULL} if no feature information is needed}

\item{featuresLoader}{a function accepting two parameters,
\code{featureFolder} and \code{components} and returning a named list of
features, or \code{NULL} if no features are needed. see
\code{\link{features.defaultLoader}} for documentation}

\item{nameProcessor}{a function which receives a relativized path to the
folder with the current model(s) and returns a name for the models}

\item{check.directory}{a function which can choose if a directory should be
followed or not}

\item{cores}{the number of cores to be used for loading}

\item{logging}{should progress information be printed: either \code{TRUE} for
printing to the console via \code{\link{print}}, \code{FALSE} for no
logging, or a path to a file receiving logging information}
}
\value{
a list of \code{\link{RegressionResults}} instances
}
\description{
Load all the setup results built by
  \code{\link{regressoR.batchLearn}} from a folder structure.
}
