% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeModels.R
\name{processMineR.mergeModels}
\alias{processMineR.mergeModels}
\title{Create a Combination or Selection of Models}
\usage{
processMineR.mergeModels(models, features.ignore = NULL,
  features.keep = NULL, namer = processMineR.joinNames)
}
\arguments{
\item{models}{the list of instances of \code{\link{RegressionResults}}}

\item{features.ignore}{a vector of names of features to delete, or
\code{NULL} if no features should explicitly be deleted}

\item{features.keep}{a vector of the names of the (only) features to keep, or
\code{NULL} if all non-deleted features should be kept}

\item{namer}{a function used to join names of model sets that now have
identical features, see \code{\link{processMineR.joinNames}} for
documentation}
}
\value{
a list of \code{\link{Models}} where all names have been transformed
  according to \code{namer} and the features have been selected according to
  \code{features.ignore} and \code{features.keep} where the \code{models}
  members of instances with the same features have been merged
}
\description{
Merge a set of models according to a given name transformation.
}
