% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPlot.R
\name{batchPlot.Models}
\alias{batchPlot.Models}
\title{Plot all the Curves from a List of Models}
\usage{
batchPlot.Models(models, log = "", plotPoints = FALSE, plotFun = TRUE,
  printNames = (length(models) < 10))
}
\arguments{
\item{models}{an instance or list of \code{\link{Models}}}

\item{log}{the logarithmic scaling info for the axes}

\item{plotPoints}{should the points be plotted?}

\item{plotFun}{should the fitted function models be plotted?}

\item{printNames}{should the model names be printed as legend}
}
\description{
A simple utility method for visualizing the data.
}
